+++ src/composer/Bootstrap.php
@@ -4,19 +4,15 @@ namespace davidhirtz\yii2\cms\hotspot\composer;
 
 use davidhirtz\yii2\cms\models\Asset;
 use davidhirtz\yii2\cms\models\ModelCloneEvent;
-use davidhirtz\yii2\cms\hotspot\assets\AdminAsset;
 use davidhirtz\yii2\cms\hotspot\models\Hotspot;
-use davidhirtz\yii2\cms\modules\admin\widgets\forms\AssetActiveForm;
 use davidhirtz\yii2\cms\hotspot\models\HotspotAsset;
 use davidhirtz\yii2\cms\hotspot\modules\admin\Module;
 use davidhirtz\yii2\cms\modules\admin\widgets\grids\columns\AssetThumbnailColumn;
+use davidhirtz\yii2\media\modules\admin\widgets\forms\fields\AssetPreview;
 use davidhirtz\yii2\skeleton\web\Application;
 use yii\base\BootstrapInterface;
 use Yii;
 use yii\base\ModelEvent;
-use yii\base\WidgetEvent;
-use yii\helpers\Json;
-use yii\helpers\Url;
 use yii\i18n\PhpMessageSource;
 
 class Bootstrap implements BootstrapInterface
@@ -52,48 +48,10 @@ class Bootstrap implements BootstrapInterface
             ],
         ]);
 
-        // Overrides asset thumbnail column.
-        $definition = \davidhirtz\yii2\cms\hotspot\modules\admin\widgets\grids\columns\AssetThumbnailColumn::class;
-        Yii::$container->set(AssetThumbnailColumn::class, $definition);
-
-        // Registers javascript after `AssetActiveForm` is rendered
-        if (!Yii::$app->getRequest()->getIsConsoleRequest()) {
-            Yii::$container->set(Asset::instance()->getActiveForm(), [
-                'on afterRun' => function (WidgetEvent $event) {
-                    /** @var AssetActiveForm $form */
-                    $form = $event->sender;
-                    /** @var Module $module */
-                    $module = Yii::$app->getModule('admin')->getModule('hotspot');
-
-                    if ($form->model->isSectionAsset() ? $module->enableSectionAssetHotspots : $module->enableEntryAssetHotspots) {
-                        if ($form->model->file->hasPreview()) {
-                            if (!$form->model->isRelationPopulated('hotspots')) {
-                                if ($form->model->getAttribute('hotspot_count')) {
-                                    $hotspots = Hotspot::find()
-                                        ->where(['asset_id' => $form->model->id])
-                                        ->orderBy(['position' => SORT_ASC])
-                                        ->all();
-                                }
-
-                                $form->model->populateRelation('hotspots', $hotspots ?? []);
-                            }
-
-                            $hotspots = $form->model->getRelatedRecords()['hotspots'] ?? [];
-
-                            $options = array_filter([
-                                'formName' => Hotspot::instance()->formName(),
-                                'url' => Url::toRoute(['/admin/hotspot/create', 'id' => $form->model->id]),
-                                'message' => !$hotspots ? Yii::t('hotspot', 'Double click on the image to create a hotspot.') : null,
-                                'hotspots' => $hotspots,
-                            ]);
-
-                            AdminAsset::register($view = $form->getView());
-                            $view->registerJs('Skeleton.registerHotspots(' . Json::htmlEncode($options) . ')');
-                        }
-                    }
-                }
-            ]);
-        }
+        Yii::$container->setDefinitions([
+            AssetThumbnailColumn::class => \davidhirtz\yii2\cms\hotspot\modules\admin\widgets\grids\columns\AssetThumbnailColumn::class,
+            AssetPreview::class => \davidhirtz\yii2\cms\hotspot\modules\admin\widgets\forms\fields\AssetPreview::class,
+        ]);
 
         // Makes sure hotspots (and their assets) are deleted on asset deletion
         ModelEvent::on(Asset::class, Asset::EVENT_BEFORE_DELETE, function (ModelEvent $event) {