+++ src/modules/admin/views/hotspot-asset/index.php
@@ -10,7 +10,7 @@
 
 use davidhirtz\yii2\cms\modules\admin\widgets\navs\Submenu;
 use davidhirtz\yii2\cms\hotspot\models\Hotspot;
-use davidhirtz\yii2\media\modules\admin\widgets\grid\FileGridView;
+use davidhirtz\yii2\media\modules\admin\widgets\grids\FileGridView;
 use davidhirtz\yii2\skeleton\web\View;
 use davidhirtz\yii2\skeleton\widgets\bootstrap\Panel;
 use yii\data\ActiveDataProvider;