+++ CHANGELOG.md
@@ -5,4 +5,8 @@ CHANGELOG
 
 - Moved source code to `src` folder
 - Moved models and widgets out of `base` folder, to override them use Yii's dependency injection
-    container
+  container
+- Changed namespaces from `davidhirtz\yii2\cms\hotspot\admin\widgets\grid`
+  to `davidhirtz\yii2\cms\hotspot\admin\widgets\grids` and `davidhirtz\yii2\cms\hotspot\admin\widgets\nav`
+  to `davidhirtz\yii2\cms\hotspot\admin\widgets\navs`
+- Added `AssetPreview` as an improved replacement for the default asset preview